requirements.txt
pygameconda install pip --force-reinstall